+++ tests/Command/Trait/InstancesTraitTest.php
@@ -16,13 +16,32 @@ use PrecisionSoft\Symfony\Phpunit\MockDto;
 use PrecisionSoft\Symfony\Phpunit\TestCase\AbstractTestCase;
 use ReflectionMethod;
 use ReflectionProperty;
+use Symfony\Component\Console\Attribute\AsCommand;
+use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\OutputInterface;
 
 class InstancesTraitTestObject
 {
     use InstancesTrait;
 
     public InputInterface $input;
+}
+
+#[AsCommand(name: 'test:instances-trait')]
+class InstancesTraitConfigureTestCommand extends Command
+{
+    use InstancesTrait;
+
+    protected function configure(): void
+    {
+        $this->configureInstances();
+    }
+
+    protected function execute(InputInterface $inputInterface, OutputInterface $outputInterface): int
+    {
+        return self::SUCCESS;
+    }
 }
 
 /**
@@ -162,6 +181,19 @@ final class InstancesTraitTest extends AbstractTestCase
         $computedInstances = $reflectionMethod->invoke($traitObject);
 
         static::assertSame([3, 3], $computedInstances);
+    }
+
+    public function testConfigureInstancesRegistersOptions(): void
+    {
+        $instancesTraitConfigureTestCommand = new InstancesTraitConfigureTestCommand();
+
+        $maxInstancesDefinition = $instancesTraitConfigureTestCommand->getDefinition()->getOption('max-instances');
+        $instanceIndexDefinition = $instancesTraitConfigureTestCommand->getDefinition()->getOption('instance-index');
+
+        static::assertTrue($maxInstancesDefinition->isValueOptional());
+        static::assertSame(1, $maxInstancesDefinition->getDefault());
+        static::assertTrue($instanceIndexDefinition->isValueOptional());
+        static::assertSame(1, $instanceIndexDefinition->getDefault());
     }
 
     public function testComputeInstancesSingleInstance(): void